+++ R/LearnerTabNet.R
@@ -13,13 +13,8 @@
 #' mlr3::mlr_learners$get("classif.tabnet")
 #' mlr3::lrn("classif.tabnet")
 #' ```
-#' 
+#' @template tabnet_description 
 #' @references Sercan, A. and Pfister, T. (2019): TabNet. \url{https://arxiv.org/abs/1908.07442}.
-#' @description
-#' Most of the parameters can be obtained from the paper.
-#' Some exceptions are documented here.
-#' * `output_dim`: Dimensions of the pen-ultimate layer(s).
-#' * `feature_dim`: Dimension of the intermediate feature representations.
 #' @template seealso_learner
 #' @templateVar learner_name classif.tabnet
 #' @template example
@@ -29,28 +24,111 @@ LearnerClassifTabNet = R6::R6Class("LearnerClassifTabNet",
   public = list(
     initialize = function() {
       ps = ParamSet$new(list(
+        ParamLgl$new("stacked", default = FALSE, tags = "train"),
+        ParamInt$new("num_layers", lower = 1, upper = Inf, default = 1L, tags = "train"),
         ParamDbl$new("batch_momentum", lower = 0, upper = 1, tags = "train"),
-        ParamDbl$new("relaxation_factor", lower = 0, upper = Inf, default = 1, tags = "train"),
+        ParamDbl$new("relaxation_factor", lower = 0, upper = Inf, default = 1.5, tags = "train"),
         ParamDbl$new("sparsity_coefficient", lower = 0, upper = 1, default = 10^-5, tags = "train"),
-        ParamInt$new("num_decision_steps", lower = 1L, upper = Inf, default = 2L, tags = "train"),
-        ParamInt$new("feature_dim", lower = 1L, upper = Inf, default = 4L, tags = "train"),
-        ParamInt$new("output_dim", lower = 1L, upper = Inf, default = 4L, tags = "train"),
+        ParamInt$new("num_decision_steps", lower = 1L, upper = Inf, default = 5L, tags = "train"),
+        ParamInt$new("feature_dim", lower = 1L, upper = Inf, default = 64L, tags = "train"),
+        ParamInt$new("output_dim", lower = 1L, upper = Inf, default = 64L, tags = "train"),
+        ParamInt$new("num_groups", lower = 1L, upper = Inf, default = 2L, tags = "train"),
+        ParamDbl$new("epsilon", lower = 0, upper = 1, default = 10^-5, tags = "train"),
+        ParamFct$new("norm_type", levels = c("group", "batch"), default = "group", tags = "train"),
+        ParamInt$new("virtual_batch_size", lower = 1L, upper = Inf, tags = "train", special_vals = list(NULL)),
         ParamFct$new("loss", default = "categorical_crossentropy", tags = "train",
           levels = c("categorical_crossentropy", "sparse_categorical_crossentropy")),
         ParamUty$new("optimizer", default = "optimizer_adam(3*10^-4)", tags = "train"),
         ParamUty$new("metrics", default = "accuracy", tags = "train")
       ))
+      ps$add_dep("num_layers", "stacked", CondEqual$new(TRUE))
       ps$values = list(
+        stacked = FALSE,
         batch_momentum = 0.98,
         relaxation_factor = 1.0,
         sparsity_coefficient = 10^-5,
         num_decision_steps = 2L,
         output_dim = 4L,
         feature_dim = 4L,
+        epsilon = 10^-5,
+        norm_type = "group",
+        num_groups = 2L,
+        virtual_batch_size = NULL,
         optimizer = optimizer_adam(lr = 3*10^-4),
         loss = "categorical_crossentropy",
         metrics = "accuracy"
       )
+
+      arch = KerasArchitectureTabNet$new(build_arch_fn = build_keras_tabnet, param_set = ps)
+      super$initialize(architecture = arch)
+      self$param_set$values$validation_split = 0 # Does not to work with tf_data.
+    }
+  )
+)
+
+
+#' @title Keras TabNet Neural Network for Regression
+#' 
+#' Implementation of "TabNet" from the paper TabNet: Attentive Interpretable Tabular Learning (Sercan, Pfister, 2019).
+#' See https://arxiv.org/abs/1908.07442 for details.
+#'
+#' @usage NULL
+#' @aliases mlr_learners_regr.tabnet
+#' @format [R6::R6Class()] inheriting from [mlr3keras::LearnerRegrKeras].
+#'
+#' @section Construction:
+#' ```
+#' LearnerRegrTabNet$new()
+#' mlr3::mlr_learners$get("regr.tabnet")
+#' mlr3::lrn("regr.tabnet")
+#' ```
+#' @template tabnet_description 
+#' @references Sercan, A. and Pfister, T. (2019): TabNet. \url{https://arxiv.org/abs/1908.07442}.
+#' @template seealso_learner
+#' @templateVar learner_name regr.tabnet
+#' @template example
+#' @export
+LearnerRegrTabNet = R6::R6Class("LearnerRegrTabNet",
+  inherit = LearnerRegrKeras,
+  public = list(
+    initialize = function() {
+      ps = ParamSet$new(list(
+        ParamLgl$new("stacked", default = FALSE, tags = "train"),
+        ParamInt$new("num_layers", lower = 1, upper = Inf, default = 1L, tags = "train"),
+        ParamDbl$new("batch_momentum", lower = 0, upper = 1, tags = "train"),
+        ParamDbl$new("relaxation_factor", lower = 0, upper = Inf, default = 1.5, tags = "train"),
+        ParamDbl$new("sparsity_coefficient", lower = 0, upper = 1, default = 10^-5, tags = "train"),
+        ParamInt$new("num_decision_steps", lower = 1L, upper = Inf, default = 5L, tags = "train"),
+        ParamInt$new("feature_dim", lower = 1L, upper = Inf, default = 64L, tags = "train"),
+        ParamInt$new("output_dim", lower = 1L, upper = Inf, default = 64L, tags = "train"),
+        ParamInt$new("num_groups", lower = 1L, upper = Inf, default = 2L, tags = "train"),
+        ParamDbl$new("epsilon", lower = 0, upper = 1, default = 10^-5, tags = "train"),
+        ParamFct$new("norm_type", levels = c("group", "batch"), default = "group", tags = "train"),
+        ParamInt$new("virtual_batch_size", lower = 1L, upper = Inf, tags = "train", special_vals = list(NULL)),
+        ParamUty$new("optimizer", default = "optimizer_adam(3*10^-4)", tags = "train"),
+        ParamFct$new("loss", default = "mean_squared_error", tags = "train",
+          levels = c("cosine_proximity", "cosine_similarity", "mean_absolute_error", "mean_squared_error",
+            "poison", "squared_hinge", "mean_squared_logarithmic_error")),
+        ParamUty$new("metrics", default = "mean_squared_logarithmic_error", tags = "train")
+      ))
+      ps$add_dep("num_layers", "stacked", CondEqual$new(TRUE))
+      ps$values = list(
+        stacked = FALSE,
+        batch_momentum = 0.98,
+        relaxation_factor = 1.0,
+        sparsity_coefficient = 10^-5,
+        num_decision_steps = 2L,
+        output_dim = 4L,
+        feature_dim = 4L,
+        epsilon = 10^-5,
+        norm_type = "group",
+        num_groups = 2L,
+        virtual_batch_size = NULL,
+        optimizer = optimizer_adam(lr = 3*10^-4),
+        loss = "mean_squared_error",
+        metrics = "mean_squared_logarithmic_error"
+      )
+
       arch = KerasArchitectureTabNet$new(build_arch_fn = build_keras_tabnet, param_set = ps)
       super$initialize(architecture = arch)
       self$param_set$values$validation_split = 0 # Does not to work with tf_data.
@@ -67,7 +145,7 @@ KerasArchitectureTabNet = R6::R6Class("KerasArchitectureTabNet",
   public = list(
     initialize = function(build_arch_fn, x_transform, y_transform, param_set) {
       x_transform = function(task, pars) {
-        x = lapply(task$feature_names, function(x) { as.matrix(task$data(cols = x))})
+        x = lapply(task$feature_names, function(x) {as.matrix(task$data(cols = x))})
         names(x) = task$feature_names
         return(x)
       }
@@ -84,18 +162,26 @@ build_keras_tabnet = function(task, pars) {
     stop("Python module tabnet is not available. In order to install it use 
       keras::install_keras(extra_packages = c('tensorflow-hub', 'tabnet==0.1.4.1').")
   }
-  
   tabnet = reticulate::import("tabnet")
-  feature_columns = make_tf_feature_cols(task)
 
-  model = tabnet$TabNetClassifier(feature_columns, num_classes = length(task$class_names),
-    feature_dim = pars$feature_dim, output_dim = pars$output_dim, 
-    num_decision_steps = pars$num_decision_steps,
-    relaxation_factor = pars$relaxation_factor,
-    sparsity_coefficient= pars$sparsity_coefficient,
-    batch_momentum = pars$batch_momentum,
-    # FIXME: Understand / Read more about the parameters below.
-    virtual_batch_size = NULL, norm_type = 'group', num_groups = 1)
+  feature_columns = make_tf_feature_cols(task)
+  tabnet_param_names = c("feature_dim", "output_dim", "num_decision_steps", "relaxation_factor", 
+    "sparsity_coefficient", "virtual_batch_size", "norm_type", "num_groups")
+  if (pars$stacked) tabnet_param_names = c(tabnet_param_names, "num_layers")
+
+  if (inherits(task, "TaskClassif")) {
+    if (pars$stacked) clf = tabnet$StackedTabNetClassifier
+    else clf = tabnet$TabNetClassifier
+    model = invoke(clf, 
+      feature_columns = feature_columns, num_classes = length(task$class_names),
+      .args = pars[tabnet_param_names])
+  } else if (inherits(task, "TaskRegr")) {
+    if (pars$stacked) regr = tabnet$StackedTabNetRegressor
+    else regr = tabnet$TabNetRegressor
+    model = invoke(regr, 
+      feature_columns = feature_columns, num_regressors = 1L,
+      .args = pars[tabnet_param_names])
+  }
 
   model %>% compile(
     loss = pars$loss,