+++ tests/testthat/test_tabnet.R
@@ -9,3 +9,25 @@ test_that("autotest classif tabnet", {
   expect_true(result, info = result$error)
   k_clear_session()
 })
+
+test_that("autotest classif stacked tabnet", {
+  skip_if_not(reticulate::py_module_available("tabnet"))
+  learner = LearnerClassifTabNet$new()
+  learner$param_set$values$epochs = 3L
+  learner$param_set$values$stacked = TRUE
+  learner$param_set$values$num_layers = 2L
+  expect_learner(learner)
+  result = run_autotest(learner, exclude = "(feat_single|sanity)")
+  expect_true(result, info = result$error)
+  k_clear_session()
+})
+
+test_that("autotest regr tabnet", {
+  skip_if_not(reticulate::py_module_available("tabnet"))
+  learner = LearnerRegrTabNet$new()
+  learner$param_set$values$epochs = 3L
+  expect_learner(learner)
+  result = run_autotest(learner, exclude = "(feat_single|sanity)")
+  expect_true(result, info = result$error)
+  k_clear_session()
+})